Finished day 3 part 1
I learned that care must be taken with objects in list. I commented
in the code about my issue, that when using a for loop on a list,
the objects given must be copies, (which I misunderstood no pointers
to mean, all ways of accessing something are differnt)

Update: per this article:
 https://stackoverflow.com/questions/
988155/how-do-i-operate-on-the-actual-object-not-a-copy-in-a-
python-for-loop
I've found that they aren't copies, but are immutable. Duh!